requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=0.1.0
langchain-google-vertexai>=0.3.0
langchain-groq>=0.1.0
langmem>=0.0.29
langsmith>=0.1.0
openevals>=0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
jsonschema>=4.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.23.0
streamlit>=1.28.0
plotly
schedule>=1.2.0

# FastAPI and API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Additional utilities
httpx>=0.25.0
requests>=2.31.0
redis>=5.0.0
celery[redis]>=5.3.0
psutil>=5.9.0